# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HDUD API Core
  version: 0.7.0
  description: >
    HDUD Core API (v0.7) — Auth + Authors + Memories (versioning, ledger, rollback, timeline).
    Unicode-safe end-to-end (NVARCHAR + mssql NVarChar). All protected endpoints use Bearer JWT.
servers:
  - url: http://localhost:4000
    description: Local Dev
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AuthSignupRequest:
      type: object
      properties:
        full_name: { type: string, example: Alexandre Neves }
        email: { type: string, format: email, example: dba.alexandre.neves@gmail.com }
        password: { type: string, example: SenhaForte#2025 }
        author_code: { type: string, example: ALE-0001 }
      required: [full_name, email, password]
    AuthLoginRequest:
      type: object
      properties:
        email: { type: string, format: email, example: dba.alexandre.neves@gmail.com }
        password: { type: string, example: SenhaForte#2025 }
      required: [email, password]
    AuthUser:
      type: object
      properties:
        user_id: { type: integer, example: 4 }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
        full_name: { type: string, nullable: true }
        author_id: { type: string, nullable: true, example: "1" }
        roles:
          type: array
          items: { type: string, example: AUTHOR_SELF }
    AuthSignupResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
        author:
          type: object
          properties:
            author_id: { type: integer, example: 1 }
            author_code: { type: string, example: ALE-0001 }
            full_name: { type: string }
            created_at: { type: string, format: date-time }
        token: { type: string }
        access_token: { type: string }
        refresh_token: { type: string }
      required: [user, token]
    AuthLoginResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
        token: { type: string }
        access_token: { type: string }
        refresh_token: { type: string }
      required: [user, token]
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Sign up (creates user + optionally links/creates author)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthSignupRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSignupResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login (returns JWT tokens)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthLoginResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...